Use catalog for mock data.
