+++ data/local/src/test/java/com/tfandkusu/template/data/local/GithubRepoLocalDataStoreTest.kt
@@ -2,9 +2,9 @@ package com.tfandkusu.template.data.local
 
 import android.content.Context
 import androidx.test.core.app.ApplicationProvider
+import com.tfandkusu.template.catalog.GitHubRepoCatalog
 import com.tfandkusu.template.data.local.db.AppDatabaseBuilder
 import com.tfandkusu.template.data.local.entity.LocalCreated
-import com.tfandkusu.template.model.GithubRepo
 import com.tfandkusu.template.util.CurrentTimeGetter
 import com.tfandkusu.template.util.parseUTC
 import io.kotlintest.shouldBe
@@ -49,45 +49,16 @@ class GithubRepoLocalDataStoreTest {
         every {
             currentTimeGetter.get()
         } returns firstTime.time andThen secondTime.time
-        val repo1 = GithubRepo(
-            229475311L,
-            "observe_room",
-            listOf(
-                "Check how to use Room to observe SQLite database ",
-                "and reflect the changes in the RecyclerView."
-            ).joinToString(separator = ""),
-            parseUTC("2021-10-29T00:15:46Z"),
-            "Kotlin",
-            "https://github.com/tfandkusu/observe_room",
-            false
-        )
-        val repo2 = GithubRepo(
-            320900929L,
-            "groupie_sticky_header_sample",
-            "Sample app for sticky header on the groupie",
-            parseUTC("2021-01-19T19:46:27Z"),
-            "Java",
-            "https://github.com/tfandkusu/groupie_sticky_header_sample",
-            false
-        )
-        val repo3 = GithubRepo(
-            343133709L,
-            "conference-app-2021",
-            "The Official App for DroidKaigi 2021",
-            parseUTC("2021-09-21T16:56:04Z"),
-            "Kotlin",
-            "https://github.com/tfandkusu/conference-app-2021",
-            true
-        )
-        localDataStore.save(listOf(repo1, repo2, repo3))
-        localDataStore.listAsFlow().first() shouldBe listOf(repo1, repo3, repo2)
+        val (repo1, repo2, repo3) = GitHubRepoCatalog.getList()
+        localDataStore.save(listOf(repo1, repo3, repo2))
+        localDataStore.listAsFlow().first() shouldBe listOf(repo1, repo2, repo3)
         createdLocalDataStore.get(LocalCreated.KIND_GITHUB_REPO) shouldBe firstTime.time
         // Second save
-        localDataStore.save(listOf(repo1.copy(name = "Edited"), repo2, repo3))
+        localDataStore.save(listOf(repo1.copy(name = "Edited"), repo3, repo2))
         localDataStore.listAsFlow().first() shouldBe listOf(
             repo1.copy(name = "Edited"),
-            repo3,
-            repo2
+            repo2,
+            repo3
         )
         createdLocalDataStore.get(LocalCreated.KIND_GITHUB_REPO) shouldBe secondTime.time
         // Check clear
