+++ data/repository/src/test/java/com/tfandkusu/template/data/repository/GithubRepoRepositoryTest.kt
@@ -1,12 +1,12 @@
 package com.tfandkusu.template.data.repository
 
+import com.tfandkusu.template.catalog.GitHubRepoCatalog
 import com.tfandkusu.template.data.local.CreatedLocalDataStore
 import com.tfandkusu.template.data.local.GithubRepoLocalDataStore
 import com.tfandkusu.template.data.local.entity.LocalCreated
 import com.tfandkusu.template.data.remote.GithubRemoteDataStore
 import com.tfandkusu.template.model.GithubRepo
 import com.tfandkusu.template.util.CurrentTimeGetter
-import com.tfandkusu.template.util.parseUTC
 import io.kotlintest.shouldBe
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
@@ -49,37 +49,7 @@ class GithubRepoRepositoryTest {
             createdLocalDataStore,
             currentTimeGetter
         )
-        val repo1 = GithubRepo(
-            229475311L,
-            "observe_room",
-            listOf(
-                "Check how to use Room to observe SQLite database ",
-                "and reflect the changes in the RecyclerView."
-            ).joinToString(separator = ""),
-            parseUTC("2021-10-29T00:15:46Z"),
-            "Kotlin",
-            "https://github.com/tfandkusu/observe_room",
-            false
-        )
-        val repo2 = GithubRepo(
-            320900929L,
-            "groupie_sticky_header_sample",
-            "Sample app for sticky header on the groupie",
-            parseUTC("2021-01-19T19:46:27Z"),
-            "Java",
-            "https://github.com/tfandkusu/groupie_sticky_header_sample",
-            false
-        )
-        val repo3 = GithubRepo(
-            343133709L,
-            "conference-app-2021",
-            "The Official App for DroidKaigi 2021",
-            parseUTC("2021-09-21T16:56:04Z"),
-            "Kotlin",
-            "https://github.com/tfandkusu/conference-app-2021",
-            true
-        )
-        this.repos = listOf(repo1, repo2, repo3)
+        this.repos = GitHubRepoCatalog.getList()
     }
 
     @Test
